fix unit tests. Separate tests rleated to jwks handler to into own file
